restaurant-finder/requirements.txt
# Core dependencies
beanis>=0.1.3
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0.0
geoalchemy2>=0.14.0
psycopg2-binary>=2.9.9
redis>=5.0.0

# Data fetching
requests>=2.31.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
